% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{classify}
\alias{classify}
\title{Classify new data based on trained models.}
\usage{
classify(data, mtype = "gbm")
}
\arguments{
\item{data}{New data with structure [].}

\item{mtype}{Type of the model to be returned. Valid value [svm | rf | gbm]. Defaults to the highest accuracy gbm model.}
}
\value{
a caret::train binary classification model object trained using specified model types.
}
\description{
Classify new customers as "PHOTOGRAPHERS" or "OTHERS" by scorin against a pre-trained persistent binary classification model.
}
